Remove catch-22 participate section
The Participate section suggested to participate in the issue tracker.
However, both the ability to open an issue and to comment
has been restricted to users that have contributed to this 
repository in the past.

It is not useful to include in the spec a participate option
which can only be fulfilled by those that already have the
inner know-how, and it is highly misleading to others,
that would be led to consider their contributions welcome, 
just to find out they are actually blocked at the point where
they go to publish them.

Since on removing the link to the Issue tracker the Participate 
section ends up empty, remove it as well.